Refactored main API test classes.
Separated unit-ish from integration-ish tests, and moved constants to
their own file.
